test/alg-yescrypt.c: Fix 'OVERFLOW_BEFORE_WIDEN' found by Covscan.
CWE-190: Unintentional integer overflow (OVERFLOW_BEFORE_WIDEN)

overflow_before_widen: Potentially overflowing expression
'r << 7UL + N_log2' with type 'uint32_t' (32 bits, unsigned) is
evaluated using 32-bit arithmetic, and then used in a context that
expects an expression of type 'uint64_t' (64 bits, unsigned).

+++ test/alg-yescrypt.c
@@ -343,7 +343,7 @@ int main(void)
 		rom_bytes = (uint64_t)r << (7 + NROM_log2);
 
 		N_log2 = 0;
-		while ((r << (7 + N_log2)) < ram_bytes)
+		while (((uint64_t)r << (7 + N_log2)) < ram_bytes)
 			N_log2++;
 		ram_bytes = (uint64_t)r << (7 + N_log2);
